requirements.txt
fastapi
uvicorn[standard]
pinecone
sentence-transformers
PyMuPDF
pillow
python-dotenv
logging
streamlit
requests
python-multipart